+++ pytorch_lightning/utilities/model_utils.py
@@ -1,8 +1,9 @@
 from pytorch_lightning.core.lightning import LightningModule
 from pytorch_lightning.core.datamodule import LightningDataModule
+from typing import Union
 
 
-def is_overridden(method_name: str, model: LightningModule) -> bool:
+def is_overridden(method_name: str, model: Union[LightningModule, LightningDataModule]) -> bool:
     # if you pass DataModule instead of None or a LightningModule, we use LightningDataModule as super
     # TODO - refector this function to accept model_name, instance, parent so it makes more sense
     super_object = LightningModule if not isinstance(model, LightningDataModule) else LightningDataModule