+++ tests/trainer/test_lr_finder.py
@@ -171,9 +171,9 @@ def test_accumulation_and_early_stopping(tmpdir):
 
     assert before_lr != after_lr, \
         'Learning rate was not altered after running learning rate finder'
-    assert len(lrfinder.results['lr']) == 100, \
+    assert len(lrfinder.results['lr']) == 99, \
         'Early stopping for learning rate finder did not work'
-    assert lrfinder._total_batch_idx == 100 * 2, \
+    assert lrfinder._total_batch_idx == 99 * 2, \
         'Accumulation parameter did not work'